move ElementPrinterHelper
