+++ src/main/java/spoon/reflect/visitor/DefaultJavaPrettyPrinter.java
@@ -110,7 +110,6 @@ import spoon.reflect.reference.CtWildcardReference;
 import spoon.reflect.visitor.PrintingContext.Writable;
 import spoon.reflect.visitor.filter.PotentialVariableDeclarationFunction;
 import spoon.reflect.visitor.printer.CommentOffset;
-import spoon.reflect.visitor.printer.ElementPrinterHelper;
 
 import java.lang.annotation.Annotation;
 import java.util.Collection;
